backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-pymongo==2.3.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
requests==2.32.5
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
python-dotenv
pymongo==4.6.1
bcrypt==4.1.2

